Restarted work on this
Work has been restarted on this repo and config as I wanted to properly
configure vim/nvim. Therefore, knowing I had this, I got into updating
it and make it more functional.

A substantial change in structure is included, now storing all files to
be copied to the home directory in fs. This may later change to be
called home or similar, as some files in the filesystem root may be
interesting to keep too. To link the files now to your system, you
should use the `link.sh` script provided in the root of the directory
(bear in mind this expects the repository to cloned into $HOME).

Additionally, the files for NuShell are added here.

Also deletes the `.wakatime/cfg` as it does not hold any important
config and so excludes it from VCS.